requirements.txt
langchain-astradb
langchain 
langchain-google-genai
datasets
pypdf
python-dotenv
flask
pandas 
setuptools

